backend/mysite/requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.38.0
django-cors-headers==2.4.0
django-phonenumber-field==2.0.1
django-rest-auth==0.9.3
django-rest-knox==3.1.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
oauthlib==2.1.0
phonenumbers==8.9.15
pipenv==2018.10.9
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
